First pass, refactoring event processors
