+++ server/models/Event.js
@@ -1,5 +1,5 @@
 // @flow
-import events from "../events";
+import { globalEventsQueue } from "../queues";
 import { DataTypes, sequelize } from "../sequelize";
 
 const Event = sequelize.define("event", {
@@ -45,13 +45,13 @@ Event.beforeCreate((event) => {
 });
 
 Event.afterCreate((event) => {
-  events.add(event, { removeOnComplete: true });
+  globalEventsQueue.add(event, { removeOnComplete: true });
 });
 
 // add can be used to send events into the event system without recording them
 // in the database / audit trail
 Event.add = (event) => {
-  events.add(Event.build(event), { removeOnComplete: true });
+  globalEventsQueue.add(Event.build(event), { removeOnComplete: true });
 };
 
 Event.ACTIVITY_EVENTS = [
