+++ server/services/backlinks.js
@@ -1,7 +1,7 @@
 // @flow
-import type { DocumentEvent, RevisionEvent } from "../events";
 import { Document, Backlink } from "../models";
 import { Op } from "../sequelize";
+import type { DocumentEvent, RevisionEvent } from "../types";
 import parseDocumentIds from "../utils/parseDocumentIds";
 import slugify from "../utils/slugify";
