+++ server/services/importer.js
@@ -3,8 +3,8 @@ import fs from "fs";
 import os from "os";
 import File from "formidable/lib/file";
 import collectionImporter from "../commands/collectionImporter";
-import type { Event } from "../events";
 import { Attachment, User } from "../models";
+import type { Event } from "../types";
 
 export default class Importer {
   async on(event: Event) {
