+++ server/services/slack.js
@@ -1,8 +1,8 @@
 // @flow
 import fetch from "fetch-with-proxy";
-import type { DocumentEvent, IntegrationEvent, Event } from "../events";
 import { Document, Integration, Collection, Team } from "../models";
 import { presentSlackAttachment } from "../presenters";
+import type { DocumentEvent, IntegrationEvent, Event } from "../types";
 
 export default class Slack {
   async on(event: Event) {
